Consolidate trace and metric declarations
